requirements.txt
appnope==0.1.0
confusable-homoglyphs==2.0.2
decorator==4.1.2
Django==1.11.6
django-debug-toolbar==1.8
django-extensions==1.9.1
django-registration==2.3
flake8==3.4.1
hashids==1.2.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mccabe==0.6.1
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
